Округлять секунды в AddSeconds как decimal без приведения к double
0.00015m в double становится 0.00014999..., из-за чего AwayFromZero
округлял значение до 0.0001 вместо 0.0002. Теперь округление и расчёт
тиков выполняются в decimal. Добавлены тесты на граничные ±0.00015.

+++ src/OneScript.Core/Values/BslDateValue.cs
@@ -34,7 +34,7 @@ namespace OneScript.Values
 
         public static DateTime AddSeconds(DateTime date, decimal seconds)
         {
-            var rounded = Math.Round((double)seconds, 4, MidpointRounding.AwayFromZero);
+            var rounded = Math.Round(seconds, 4, MidpointRounding.AwayFromZero);
             return Normalize(date.AddTicks((long)(rounded * TimeSpan.TicksPerSecond)));
         }
